PAMI requires that all structures passed along have a size that is a multiple of 4 bytes (this is not documented, it's an implementation quirk).  The 2-byte x10rt_msg_type, as the only 2-byte value in several structures, causes these structures to violate the 4-byte rule.  Generally it works fine, but PAMI grabs two bytes beyond the structures, so there's a potential for segfaults.  Changing the x10rt_msg_type to 4 bytes corrects this, and it is a better solution than adding in a 2->4 byte adjustment with every message in our PAMI layer, as that would add an extra few steps in our most performance-critical backend.
git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@23458 a326200e-df2c-42e0-81fa-597e909af41d

+++ x10.runtime/x10rt/include/x10rt_types.h
@@ -40,7 +40,7 @@ typedef void x10rt_completion_handler2 (x10rt_team, void *arg);
 
 /** An integer type capable of representing any message type id.
  */
-typedef uint16_t x10rt_msg_type;
+typedef uint32_t x10rt_msg_type;
 
 /** An integer type capable of representing a remote void* or size_t.
  */
